Move logic from CrosswordCell to the useCrosswordCell custom hook
